apply PR feedback

+++ unity-renderer/Assets/Scripts/MainScripts/DCL/CRDTProtocol/CRDTUtils.cs
@@ -4,7 +4,10 @@ namespace DCL.CRDT
     {
         public static long KeyFromIds(int entityId, int componentId)
         {
-            return (long)entityId << 32 | (uint)componentId;
+            // we create an int64 that store both entityId and componentId.
+            // entityId on the left 32 bit and componentId in the remaining 
+            // 32 bits of the right
+            return (long)entityId << 32 | (long)componentId;
         }
 
         public static int EntityIdFromKey(long key)
